refactor: split host runtime egress pipeline
