+++ crates/ironclaw_host_runtime/src/lib.rs
@@ -26,27 +26,17 @@
 
 use async_trait::async_trait;
 use ironclaw_host_api::{
-    ApprovalRequestId, CapabilityId, CorrelationId, ExecutionContext, ExtensionId, NetworkPolicy,
-    ProcessId, ResourceEstimate, ResourceScope, ResourceUsage, RuntimeKind, SecretHandle,
-};
-use ironclaw_host_api::{
-    RuntimeCredentialInjection, RuntimeCredentialSource, RuntimeCredentialTarget,
-    RuntimeHttpEgress, RuntimeHttpEgressError, RuntimeHttpEgressRequest, RuntimeHttpEgressResponse,
-    is_sensitive_runtime_request_header, is_sensitive_runtime_response_header,
+    ApprovalRequestId, CapabilityId, CorrelationId, ExecutionContext, ExtensionId, ProcessId,
+    ResourceEstimate, ResourceScope, ResourceUsage, RuntimeKind, SecretHandle,
     runtime_policy::{DeploymentMode, EffectiveRuntimePolicy, RuntimeProfile},
 };
-use ironclaw_network::{
-    NetworkHttpEgress, NetworkHttpError, NetworkHttpRequest, NetworkHttpResponse,
-};
-use ironclaw_safety::{LeakDetector, params_contain_manual_credentials};
-use ironclaw_secrets::{SecretMaterial, SecretStore, SecretStoreError};
 use ironclaw_trust::TrustDecision;
-use secrecy::ExposeSecret;
 use serde_json::Value;
-use std::{collections::BTreeMap, env, fmt, sync::Arc};
+use std::{collections::BTreeMap, env, fmt};
 use thiserror::Error;
 
 mod capability_catalog;
+mod egress;
 mod extension_contracts;
 mod first_party;
 mod first_party_tools;
@@ -68,6 +58,7 @@ pub use capability_catalog::{
     HotCapabilityCatalog, HotCapabilityRecord, MAX_HOT_PROMPT_BYTES, MAX_HOT_SCHEMA_BYTES,
     publish_hot_capability_catalog,
 };
+pub use egress::HostHttpEgressService;
 pub use extension_contracts::{
     default_host_api_contract_registry, default_host_port_catalog,
     discover_extensions_with_default_host_api_contracts,
@@ -94,7 +85,6 @@ pub use obligations::{
     BuiltinObligationHandler, BuiltinObligationServices, LEAK_REDACT_FAILED_CODE,
     ProcessObligationLifecycleStore,
 };
-use obligations::{NetworkObligationPolicyStore, RuntimeSecretInjectionStore};
 pub use planner::{ExecutionPlan, PlannerError, plan_capability};
 pub use process_port::{
     CommandExecutionOutput, CommandExecutionRequest, LocalHostProcessPort, RuntimeProcessError,
@@ -956,892 +946,3 @@ impl HostRuntimeError {
         }
     }
 }
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum NetworkPolicySource {
-    StagedObligation,
-    RequestPolicyFallback,
-}
-
-#[derive(Debug, Clone)]
-pub struct HostHttpEgressService<N, S> {
-    network: N,
-    secrets: S,
-    secret_injections: Option<Arc<RuntimeSecretInjectionStore>>,
-    network_policy_store: Option<Arc<NetworkObligationPolicyStore>>,
-    network_policy_source: NetworkPolicySource,
-    unsafe_raw_diagnostics_allowed: bool,
-    body_store: Option<Arc<dyn RuntimeHttpBodyStore>>,
-}
-
-impl<N, S> HostHttpEgressService<N, S> {
-    /// Construct host HTTP egress in production fail-closed mode.
-    ///
-    /// Host-runtime composition must attach a staged network policy store
-    /// before executing network requests.
-    /// Without that store, egress fails before transport rather than trusting a
-    /// caller-supplied policy.
-    pub fn new(network: N, secrets: S) -> Self {
-        Self {
-            network,
-            secrets,
-            secret_injections: None,
-            network_policy_store: None,
-            network_policy_source: NetworkPolicySource::StagedObligation,
-            unsafe_raw_diagnostics_allowed: false,
-            body_store: None,
-        }
-    }
-
-    /// Construct host HTTP egress that uses the policy embedded in each request.
-    ///
-    /// This is intentionally named as a test/legacy seam: production Reborn
-    /// runtime egress must consume staged `ApplyNetworkPolicy` handoffs from
-    /// the staged network-policy store instead of trusting runtime/caller
-    /// request policy fields.
-    pub fn new_with_request_policy_for_tests(network: N, secrets: S) -> Self {
-        Self {
-            network,
-            secrets,
-            secret_injections: None,
-            network_policy_store: None,
-            network_policy_source: NetworkPolicySource::RequestPolicyFallback,
-            unsafe_raw_diagnostics_allowed: false,
-            body_store: None,
-        }
-    }
-
-    pub(crate) fn with_unsafe_raw_diagnostics_allowed(mut self, allowed: bool) -> Self {
-        self.unsafe_raw_diagnostics_allowed = allowed;
-        self
-    }
-
-    pub(crate) fn with_secret_injection_store(
-        mut self,
-        store: Arc<RuntimeSecretInjectionStore>,
-    ) -> Self {
-        self.secret_injections = Some(store);
-        self
-    }
-
-    pub(crate) fn with_network_policy_store(
-        mut self,
-        store: Arc<NetworkObligationPolicyStore>,
-    ) -> Self {
-        self.network_policy_store = Some(store);
-        self.network_policy_source = NetworkPolicySource::StagedObligation;
-        self
-    }
-
-    pub(crate) fn is_production_wired_with(
-        &self,
-        network_policy_store: &Arc<NetworkObligationPolicyStore>,
-        secret_injections: &Arc<RuntimeSecretInjectionStore>,
-    ) -> bool {
-        matches!(
-            self.network_policy_source,
-            NetworkPolicySource::StagedObligation
-        ) && self
-            .network_policy_store
-            .as_ref()
-            .is_some_and(|store| Arc::ptr_eq(store, network_policy_store))
-            && self
-                .secret_injections
-                .as_ref()
-                .is_some_and(|store| Arc::ptr_eq(store, secret_injections))
-    }
-
-    pub fn with_body_store(mut self, store: Arc<dyn RuntimeHttpBodyStore>) -> Self {
-        self.body_store = Some(store);
-        self
-    }
-
-    fn network_policy_for_request(
-        &self,
-        request: &mut RuntimeHttpEgressRequest,
-    ) -> Result<NetworkPolicy, RuntimeHttpEgressError> {
-        if let Some(store) = &self.network_policy_store {
-            return store
-                .get(&request.scope, &request.capability_id)
-                .ok_or_else(|| RuntimeHttpEgressError::Network {
-                    reason: "network_policy_missing".to_string(),
-                    request_bytes: 0,
-                    response_bytes: 0,
-                });
-        }
-
-        match self.network_policy_source {
-            NetworkPolicySource::StagedObligation => Err(RuntimeHttpEgressError::Network {
-                reason: "network_policy_missing".to_string(),
-                request_bytes: 0,
-                response_bytes: 0,
-            }),
-            NetworkPolicySource::RequestPolicyFallback => {
-                Ok(std::mem::take(&mut request.network_policy))
-            }
-        }
-    }
-
-    fn discard_staged_policy_for_request(&self, request: &RuntimeHttpEgressRequest) {
-        if let Some(store) = &self.network_policy_store {
-            store.discard_for_capability(&request.scope, &request.capability_id);
-        }
-    }
-
-    fn validate_credential_sources_for_request(
-        &self,
-        request: &RuntimeHttpEgressRequest,
-    ) -> Result<(), RuntimeHttpEgressError> {
-        if matches!(
-            self.network_policy_source,
-            NetworkPolicySource::RequestPolicyFallback
-        ) {
-            return Ok(());
-        }
-
-        for injection in &request.credential_injections {
-            match &injection.source {
-                RuntimeCredentialSource::SecretStoreLease => {
-                    return Err(RuntimeHttpEgressError::Credential {
-                        reason:
-                            "direct secret-store leases are unavailable for production runtime egress"
-                                .to_string(),
-                    });
-                }
-                RuntimeCredentialSource::StagedObligation { capability_id }
-                    if capability_id != &request.capability_id =>
-                {
-                    return Err(RuntimeHttpEgressError::Credential {
-                        reason: "staged credential capability does not match request capability"
-                            .to_string(),
-                    });
-                }
-                RuntimeCredentialSource::StagedObligation { .. } => {}
-            }
-        }
-
-        Ok(())
-    }
-}
-
-impl<N, S> RuntimeHttpEgress for HostHttpEgressService<N, S>
-where
-    N: NetworkHttpEgress,
-    S: SecretStore,
-{
-    fn execute(
-        &self,
-        mut request: RuntimeHttpEgressRequest,
-    ) -> Result<RuntimeHttpEgressResponse, RuntimeHttpEgressError> {
-        let network_policy = self.network_policy_for_request(&mut request)?;
-        if let Err(error) = self.validate_credential_sources_for_request(&request) {
-            self.discard_staged_policy_for_request(&request);
-            return Err(error);
-        }
-        let save_body_to = std::mem::take(&mut request.save_body_to);
-        let body_store = self.body_store.as_deref();
-        if let Some(target) = &save_body_to {
-            let Some(store) = body_store else {
-                self.discard_staged_policy_for_request(&request);
-                return Err(RuntimeHttpEgressError::Request {
-                    reason: "response_body_store_unavailable".to_string(),
-                    request_bytes: 0,
-                    response_bytes: 0,
-                });
-            };
-            if let Err(error) =
-                store.authorize_write(&request.scope, &request.capability_id, target)
-            {
-                self.discard_staged_policy_for_request(&request);
-                return Err(RuntimeHttpEgressError::Request {
-                    reason: format!("response_body_store_unauthorized: {}", error.reason),
-                    request_bytes: 0,
-                    response_bytes: 0,
-                });
-            }
-        }
-        if let Err(error) = validate_runtime_request(&request) {
-            self.discard_staged_policy_for_request(&request);
-            return Err(error);
-        }
-        let scope = request.scope.clone();
-        let capability_id = request.capability_id.clone();
-
-        let mut redaction_values = Vec::new();
-        let mut credential_materials = Vec::new();
-        let mut parsed_url = None;
-        let credential_injections = std::mem::take(&mut request.credential_injections);
-        for injection in &credential_injections {
-            let value = match credential_value_for_injection(
-                &mut credential_materials,
-                &self.secrets,
-                self.secret_injections.as_deref(),
-                &request,
-                injection,
-            ) {
-                Ok(value) => value,
-                Err(error) => {
-                    self.discard_staged_policy_for_request(&request);
-                    return Err(error);
-                }
-            };
-            let Some(value) = value else {
-                continue;
-            };
-            // Borrow the leased plaintext only for the narrow window where the
-            // egress code needs it (header/query injection + redaction-token
-            // generation). The `SecretMaterial` stays inside the cache; the
-            // exposed `&str` does not outlive this loop iteration. Plaintext
-            // does land in `request.headers` and `redaction_values` after
-            // injection — those copies are unavoidable because the network
-            // layer and response-body scanner consume raw bytes — but the
-            // cache itself never holds a non-zeroizing copy.
-            let plaintext = value.expose_secret();
-            if let Err(error) = apply_credential_injection(
-                &mut request,
-                &mut parsed_url,
-                &injection.target,
-                plaintext,
-            ) {
-                self.discard_staged_policy_for_request(&request);
-                return Err(error);
-            }
-            redaction_values.extend(redaction_values_for_secret(plaintext));
-        }
-        if let Some(url) = parsed_url {
-            request.url = url.to_string();
-        }
-
-        let response = self
-            .network
-            .execute(NetworkHttpRequest {
-                scope: request.scope,
-                method: request.method,
-                url: request.url,
-                headers: request.headers,
-                body: request.body,
-                policy: network_policy,
-                response_body_limit: request.response_body_limit,
-                timeout_ms: request.timeout_ms,
-            })
-            .map_err(|error| runtime_network_error(self.unsafe_raw_diagnostics_allowed, error))?;
-        let credentials_injected = !redaction_values.is_empty();
-        let (response, response_redacted) = sanitize_runtime_response(response, &redaction_values)?;
-        let (response, saved_body) = http_body::apply_body_disposition(
-            response,
-            save_body_to,
-            body_store,
-            &scope,
-            &capability_id,
-        )?;
-        Ok(runtime_response(
-            response,
-            credentials_injected || response_redacted,
-            saved_body,
-        ))
-    }
-}
-
-struct RuntimeCredentialMaterialCacheEntry {
-    key: RuntimeCredentialMaterialKey,
-    /// Leased secret material kept inside `SecretString` so the bytes are
-    /// zeroized when this entry — and its enclosing `Vec` — is dropped at
-    /// the end of the egress call. Holding plaintext as `String` here
-    /// instead would leave the leased credential on the heap for the
-    /// duration of the request, defeating `SecretMaterial::ZeroizeOnDrop`.
-    value: Option<SecretMaterial>,
-}
-
-#[derive(Clone, PartialEq, Eq)]
-enum RuntimeCredentialMaterialKey {
-    SecretStoreLease {
-        handle: SecretHandle,
-    },
-    StagedObligation {
-        capability_id: CapabilityId,
-        handle: SecretHandle,
-    },
-}
-
-impl RuntimeCredentialMaterialKey {
-    fn for_injection(injection: &RuntimeCredentialInjection) -> Self {
-        match &injection.source {
-            RuntimeCredentialSource::SecretStoreLease => Self::SecretStoreLease {
-                handle: injection.handle.clone(),
-            },
-            RuntimeCredentialSource::StagedObligation { capability_id } => Self::StagedObligation {
-                capability_id: capability_id.clone(),
-                handle: injection.handle.clone(),
-            },
-        }
-    }
-}
-
-fn credential_value_for_injection<'cache, S>(
-    cache: &'cache mut Vec<RuntimeCredentialMaterialCacheEntry>,
-    secrets: &S,
-    secret_injections: Option<&RuntimeSecretInjectionStore>,
-    request: &RuntimeHttpEgressRequest,
-    injection: &RuntimeCredentialInjection,
-) -> Result<Option<&'cache SecretMaterial>, RuntimeHttpEgressError>
-where
-    S: SecretStore,
-{
-    let key = RuntimeCredentialMaterialKey::for_injection(injection);
-    if let Some(idx) = cache.iter().position(|entry| entry.key == key) {
-        // Negative cache hit (missing optional credential on a prior pass)
-        // must still error out if *this* injection marks the same handle as
-        // required. `required` is per-injection, not per-cache-entry.
-        if cache[idx].value.is_none() && injection.required {
-            return Err(RuntimeHttpEgressError::Credential {
-                reason: "required credential is unavailable".to_string(),
-            });
-        }
-        return Ok(cache[idx].value.as_ref());
-    }
-
-    let value = secret_material_for_injection(secrets, secret_injections, request, injection)?;
-    let pushed_index = cache.len();
-    cache.push(RuntimeCredentialMaterialCacheEntry { key, value });
-    Ok(cache[pushed_index].value.as_ref())
-}
-
-fn secret_material_for_injection<S>(
-    secrets: &S,
-    secret_injections: Option<&RuntimeSecretInjectionStore>,
-    request: &RuntimeHttpEgressRequest,
-    injection: &RuntimeCredentialInjection,
-) -> Result<Option<SecretMaterial>, RuntimeHttpEgressError>
-where
-    S: SecretStore,
-{
-    match &injection.source {
-        RuntimeCredentialSource::SecretStoreLease => {
-            lease_secret_for_injection(secrets, request, injection)
-        }
-        RuntimeCredentialSource::StagedObligation { capability_id } => {
-            take_staged_secret_for_injection(secret_injections, request, capability_id, injection)
-        }
-    }
-}
-
-fn take_staged_secret_for_injection(
-    secret_injections: Option<&RuntimeSecretInjectionStore>,
-    request: &RuntimeHttpEgressRequest,
-    capability_id: &CapabilityId,
-    injection: &RuntimeCredentialInjection,
-) -> Result<Option<SecretMaterial>, RuntimeHttpEgressError> {
-    let Some(secret_injections) = secret_injections else {
-        return missing_runtime_credential(injection.required);
-    };
-    match secret_injections.take(&request.scope, capability_id, &injection.handle) {
-        Ok(Some(material)) => Ok(Some(material)),
-        Ok(None) => missing_runtime_credential(injection.required),
-        Err(_) => Err(RuntimeHttpEgressError::Credential {
-            reason: "runtime credential injection store unavailable".to_string(),
-        }),
-    }
-}
-
-fn missing_runtime_credential(
-    required: bool,
-) -> Result<Option<SecretMaterial>, RuntimeHttpEgressError> {
-    if required {
-        Err(RuntimeHttpEgressError::Credential {
-            reason: "required credential is unavailable".to_string(),
-        })
-    } else {
-        Ok(None)
-    }
-}
-
-fn lease_secret_for_injection<S>(
-    secrets: &S,
-    request: &RuntimeHttpEgressRequest,
-    injection: &RuntimeCredentialInjection,
-) -> Result<Option<SecretMaterial>, RuntimeHttpEgressError>
-where
-    S: SecretStore,
-{
-    block_on_secret_store(async {
-        let metadata = secrets.metadata(&request.scope, &injection.handle).await?;
-        if metadata.is_none() {
-            return Ok(None);
-        }
-        let lease = secrets
-            .lease_once(&request.scope, &injection.handle)
-            .await?;
-        secrets.consume(&request.scope, lease.id).await.map(Some)
-    })?
-    .map_or_else(
-        || missing_runtime_credential(injection.required),
-        |material| Ok(Some(material)),
-    )
-}
-
-fn block_on_secret_store<T>(
-    future: impl std::future::Future<Output = Result<T, SecretStoreError>> + Send,
-) -> Result<T, RuntimeHttpEgressError>
-where
-    T: Send,
-{
-    let joined = std::thread::scope(|scope| {
-        scope
-            .spawn(move || {
-                let runtime = tokio::runtime::Builder::new_current_thread()
-                    .enable_all()
-                    .build()
-                    .map_err(|_| RuntimeHttpEgressError::Credential {
-                        reason: "secret store runtime unavailable".to_string(),
-                    })?;
-                runtime
-                    .block_on(future)
-                    .map_err(|error| RuntimeHttpEgressError::Credential {
-                        reason: sanitized_secret_error(&error),
-                    })
-            })
-            .join()
-    });
-    joined.unwrap_or_else(|_| {
-        Err(RuntimeHttpEgressError::Credential {
-            reason: "secret store worker panicked".to_string(),
-        })
-    })
-}
-
-fn sanitized_secret_error(error: &SecretStoreError) -> String {
-    match error {
-        SecretStoreError::UnknownSecret { .. } => "credential is unavailable".to_string(),
-        SecretStoreError::UnknownLease { .. } => "credential lease is unavailable".to_string(),
-        SecretStoreError::LeaseConsumed { .. } => "credential lease was already used".to_string(),
-        SecretStoreError::LeaseRevoked { .. } => "credential lease was revoked".to_string(),
-        SecretStoreError::LeaseExpired { .. } | SecretStoreError::SecretExpired => {
-            "credential expired".to_string()
-        }
-        SecretStoreError::BackendMisconfigured { .. } => {
-            "credential store is misconfigured".to_string()
-        }
-        SecretStoreError::StoreUnavailable { .. } => "credential store unavailable".to_string(),
-    }
-}
-
-fn apply_credential_injection(
-    request: &mut RuntimeHttpEgressRequest,
-    parsed_url: &mut Option<url::Url>,
-    target: &RuntimeCredentialTarget,
-    value: &str,
-) -> Result<(), RuntimeHttpEgressError> {
-    target
-        .validate_declaration()
-        .map_err(|_| RuntimeHttpEgressError::Credential {
-            reason: "credential injection target is invalid".to_string(),
-        })?;
-    match target {
-        RuntimeCredentialTarget::Header { name, prefix } => {
-            let injected = match prefix {
-                Some(prefix) => format!("{prefix}{value}"),
-                None => value.to_string(),
-            };
-            if injected.chars().any(char::is_control) {
-                return Err(RuntimeHttpEgressError::Credential {
-                    reason: "credential injection header value is invalid".to_string(),
-                });
-            }
-            request.headers.push((name.clone(), injected));
-        }
-        RuntimeCredentialTarget::QueryParam { name } => {
-            let url = parsed_request_url(&request.url, parsed_url)?;
-            url.query_pairs_mut().append_pair(name, value);
-        }
-        RuntimeCredentialTarget::PathPlaceholder { placeholder } => {
-            if !valid_path_credential_segment(placeholder) {
-                return Err(RuntimeHttpEgressError::Credential {
-                    reason: "credential injection path placeholder is invalid".to_string(),
-                });
-            }
-            if !valid_path_credential_segment(value) {
-                return Err(RuntimeHttpEgressError::Credential {
-                    reason: "credential injection path value is invalid".to_string(),
-                });
-            }
-            let url = parsed_request_url(&request.url, parsed_url)?;
-            if url.scheme() != "https" {
-                return Err(RuntimeHttpEgressError::Credential {
-                    reason: "credential injection path placeholder requires HTTPS".to_string(),
-                });
-            }
-            let Some(_) = url.path_segments() else {
-                return Err(RuntimeHttpEgressError::Credential {
-                    reason: "credential injection target URL has no path segments".to_string(),
-                });
-            };
-            let path = url.path().to_string();
-            let path = path.strip_prefix('/').unwrap_or(&path);
-            let placeholder_count = path
-                .split('/')
-                .filter(|segment| *segment == placeholder)
-                .count();
-            match placeholder_count {
-                0 => {
-                    return Err(RuntimeHttpEgressError::Credential {
-                        reason: "credential injection path placeholder was not found".to_string(),
-                    });
-                }
-                1 => {}
-                _ => {
-                    return Err(RuntimeHttpEgressError::Credential {
-                        reason: "credential injection path placeholder must appear exactly once"
-                            .to_string(),
-                    });
-                }
-            }
-            let mut rewritten_path = String::with_capacity(path.len() + value.len());
-            for (index, segment) in path.split('/').enumerate() {
-                if index > 0 {
-                    rewritten_path.push('/');
-                }
-                if segment == placeholder {
-                    rewritten_path.push_str(value);
-                } else {
-                    rewritten_path.push_str(segment);
-                }
-            }
-            url.set_path(&rewritten_path);
-        }
-    }
-    Ok(())
-}
-
-fn parsed_request_url<'a>(
-    raw_url: &str,
-    parsed_url: &'a mut Option<url::Url>,
-) -> Result<&'a mut url::Url, RuntimeHttpEgressError> {
-    if parsed_url.is_none() {
-        *parsed_url =
-            Some(
-                url::Url::parse(raw_url).map_err(|_| RuntimeHttpEgressError::Credential {
-                    reason: "credential injection target URL is invalid".to_string(),
-                })?,
-            );
-    }
-    parsed_url
-        .as_mut()
-        .ok_or_else(|| RuntimeHttpEgressError::Credential {
-            reason: "credential injection target URL is invalid".to_string(),
-        })
-}
-
-fn valid_path_credential_segment(segment: &str) -> bool {
-    !segment.is_empty()
-        && segment != "."
-        && segment != ".."
-        && segment
-            .bytes()
-            .all(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'-' | b'.' | b'_' | b'~'))
-}
-
-fn runtime_network_error(
-    unsafe_raw_diagnostics_allowed: bool,
-    error: NetworkHttpError,
-) -> RuntimeHttpEgressError {
-    log_raw_network_http_error_for_local_diagnostics(unsafe_raw_diagnostics_allowed, &error);
-    RuntimeHttpEgressError::Network {
-        reason: error.stable_reason().to_string(),
-        request_bytes: error.request_bytes(),
-        response_bytes: error.response_bytes(),
-    }
-}
-
-fn log_raw_network_http_error_for_local_diagnostics(
-    unsafe_raw_diagnostics_allowed: bool,
-    error: &NetworkHttpError,
-) {
-    if !unsafe_raw_http_diagnostics_enabled(unsafe_raw_diagnostics_allowed) {
-        return;
-    }
-
-    tracing::warn!(
-        network_error_kind = error.kind().as_str(),
-        raw_network_error = %error,
-        unsafe_raw_diagnostics = true,
-        "unsafe raw HTTP egress error diagnostic enabled"
-    );
-}
-
-fn validate_runtime_request(
-    request: &RuntimeHttpEgressRequest,
-) -> Result<(), RuntimeHttpEgressError> {
-    if let Some((name, _)) = request
-        .headers
-        .iter()
-        .find(|(name, _)| is_sensitive_runtime_request_header(name))
-    {
-        return Err(RuntimeHttpEgressError::Request {
-            reason: format!("sensitive_header_denied:{name}"),
-            request_bytes: 0,
-            response_bytes: 0,
-        });
-    }
-
-    if runtime_request_contains_manual_credentials(request) {
-        return Err(RuntimeHttpEgressError::Request {
-            reason: "manual_credentials_denied".to_string(),
-            request_bytes: 0,
-            response_bytes: 0,
-        });
-    }
-
-    let detector = LeakDetector::new();
-    detector
-        .scan_http_request(&request.url, &request.headers, Some(&request.body))
-        .map_err(|_| runtime_request_leak_error())?;
-    scan_decoded_url_for_leaks(&detector, &request.url)?;
-    Ok(())
-}
-
-fn runtime_request_contains_manual_credentials(request: &RuntimeHttpEgressRequest) -> bool {
-    let headers = request
-        .headers
-        .iter()
-        .map(|(name, value)| serde_json::json!({ "name": name, "value": value }))
-        .collect::<Vec<_>>();
-    let params = serde_json::json!({
-        "url": request.url,
-        "headers": headers,
-    });
-    params_contain_manual_credentials(&params)
-}
-
-fn scan_decoded_url_for_leaks(
-    detector: &LeakDetector,
-    raw_url: &str,
-) -> Result<(), RuntimeHttpEgressError> {
-    let Ok(parsed) = url::Url::parse(raw_url) else {
-        return Ok(());
-    };
-
-    scan_component_for_leaks(detector, parsed.path())?;
-    if let Some(query) = parsed.query() {
-        scan_component_for_leaks(detector, query)?;
-    }
-    if !parsed.username().is_empty() {
-        scan_component_for_leaks(detector, parsed.username())?;
-    }
-    if let Some(password) = parsed.password() {
-        scan_component_for_leaks(detector, password)?;
-    }
-    for (name, value) in parsed.query_pairs() {
-        detector
-            .scan_and_clean(name.as_ref())
-            .map_err(|_| runtime_request_leak_error())?;
-        detector
-            .scan_and_clean(value.as_ref())
-            .map_err(|_| runtime_request_leak_error())?;
-    }
-    Ok(())
-}
-
-fn scan_component_for_leaks(
-    detector: &LeakDetector,
-    component: &str,
-) -> Result<(), RuntimeHttpEgressError> {
-    let decoded = percent_decode_lossy(component);
-    if decoded != component {
-        detector
-            .scan_and_clean(&decoded)
-            .map_err(|_| runtime_request_leak_error())?;
-    }
-    Ok(())
-}
-
-fn percent_decode_lossy(input: &str) -> String {
-    let bytes = input.as_bytes();
-    let mut decoded = Vec::with_capacity(bytes.len());
-    let mut index = 0;
-    while index < bytes.len() {
-        if bytes[index] == b'%'
-            && index + 2 < bytes.len()
-            && let (Some(high), Some(low)) =
-                (hex_value(bytes[index + 1]), hex_value(bytes[index + 2]))
-        {
-            decoded.push((high << 4) | low);
-            index += 3;
-            continue;
-        }
-        decoded.push(bytes[index]);
-        index += 1;
-    }
-    String::from_utf8_lossy(&decoded).into_owned()
-}
-
-fn hex_value(byte: u8) -> Option<u8> {
-    match byte {
-        b'0'..=b'9' => Some(byte - b'0'),
-        b'a'..=b'f' => Some(byte - b'a' + 10),
-        b'A'..=b'F' => Some(byte - b'A' + 10),
-        _ => None,
-    }
-}
-
-fn runtime_request_leak_error() -> RuntimeHttpEgressError {
-    RuntimeHttpEgressError::Request {
-        reason: "credential_leak_blocked".to_string(),
-        request_bytes: 0,
-        response_bytes: 0,
-    }
-}
-
-fn sanitize_runtime_response(
-    response: NetworkHttpResponse,
-    redaction_values: &[String],
-) -> Result<(NetworkHttpResponse, bool), RuntimeHttpEgressError> {
-    let NetworkHttpResponse {
-        status,
-        headers,
-        body,
-        usage,
-    } = response;
-    let mut redaction_applied = false;
-    let mut sanitized_headers = Vec::new();
-    let detector = LeakDetector::new();
-
-    for (name, value) in headers {
-        if is_sensitive_runtime_response_header(&name) {
-            redaction_applied = true;
-            continue;
-        }
-        let exact_redacted = redact(value, redaction_values);
-        if exact_redacted.contains("[REDACTED]") {
-            redaction_applied = true;
-        }
-        let cleaned = detector.scan_and_clean(&exact_redacted).map_err(|_| {
-            RuntimeHttpEgressError::Response {
-                reason: "response_leak_blocked".to_string(),
-                request_bytes: usage.request_bytes,
-                response_bytes: usage.response_bytes,
-            }
-        })?;
-        if cleaned != exact_redacted {
-            redaction_applied = true;
-        }
-        sanitized_headers.push((name, cleaned));
-    }
-
-    let body_text = String::from_utf8_lossy(&body).into_owned();
-    let exact_redacted = redact(body_text, redaction_values);
-    let exact_body_redacted = exact_redacted.contains("[REDACTED]");
-    if exact_body_redacted {
-        redaction_applied = true;
-    }
-    let cleaned =
-        detector
-            .scan_and_clean(&exact_redacted)
-            .map_err(|_| RuntimeHttpEgressError::Response {
-                reason: "response_leak_blocked".to_string(),
-                request_bytes: usage.request_bytes,
-                response_bytes: usage.response_bytes,
-            })?;
-    let leak_detector_redacted = cleaned != exact_redacted;
-    if leak_detector_redacted {
-        redaction_applied = true;
-    }
-    let body = if exact_body_redacted || leak_detector_redacted {
-        cleaned.into_bytes()
-    } else {
-        body
-    };
-
-    Ok((
-        NetworkHttpResponse {
-            status,
-            headers: sanitized_headers,
-            body,
-            usage,
-        },
-        redaction_applied,
-    ))
-}
-
-fn runtime_response(
-    response: NetworkHttpResponse,
-    redaction_applied: bool,
-    saved_body: Option<ironclaw_host_api::RuntimeHttpSavedBody>,
-) -> RuntimeHttpEgressResponse {
-    RuntimeHttpEgressResponse {
-        status: response.status,
-        headers: response.headers,
-        body: response.body,
-        saved_body,
-        request_bytes: response.usage.request_bytes,
-        response_bytes: response.usage.response_bytes,
-        redaction_applied,
-    }
-}
-
-fn redact(mut text: String, values: &[String]) -> String {
-    for value in values {
-        if !value.is_empty() {
-            text = text.replace(value, "[REDACTED]");
-        }
-    }
-    text
-}
-
-fn redaction_values_for_secret(value: &str) -> Vec<String> {
-    let mut values = Vec::new();
-    push_redaction_value(&mut values, value.to_string());
-    let encoded = url::form_urlencoded::byte_serialize(value.as_bytes()).collect::<String>();
-    push_redaction_value(&mut values, encoded.clone());
-    push_redaction_value(&mut values, lowercase_percent_escapes(&encoded));
-    let plus_encoded = encoded.replace("%20", "+");
-    push_redaction_value(&mut values, plus_encoded.clone());
-    push_redaction_value(&mut values, lowercase_percent_escapes(&plus_encoded));
-    values
-}
-
-fn push_redaction_value(values: &mut Vec<String>, value: String) {
-    if !value.is_empty() && !values.iter().any(|existing| existing == &value) {
-        values.push(value);
-    }
-}
-
-fn lowercase_percent_escapes(value: &str) -> String {
-    let bytes = value.as_bytes();
-    let mut output = String::with_capacity(value.len());
-    let mut index = 0;
-    while index < bytes.len() {
-        if bytes[index] == b'%'
-            && index + 2 < bytes.len()
-            && bytes[index + 1].is_ascii_hexdigit()
-            && bytes[index + 2].is_ascii_hexdigit()
-        {
-            output.push('%');
-            output.push((bytes[index + 1] as char).to_ascii_lowercase());
-            output.push((bytes[index + 2] as char).to_ascii_lowercase());
-            index += 3;
-            continue;
-        }
-        output.push(bytes[index] as char);
-        index += 1;
-    }
-    output
-}
-
-/// **Finding H1 — compile-time regression guard.**
-///
-/// The credential material cache value field must hold a `ZeroizeOnDrop`
-/// carrier. Holding the leased plaintext as `Option<String>` (the original
-/// bug) leaves it on the heap until the cache `Vec` is dropped at end-of-call,
-/// then frees the bytes without wiping. This `const _: fn(...) = ...`
-/// references the field's inner type through a `ZeroizeOnDrop`-bounded helper,
-/// so any refactor that downgrades the field to a non-zeroizing type (e.g.
-/// plain `Option<String>`) stops the crate from compiling rather than waiting
-/// for a test run. `String` implements `Zeroize` but not `ZeroizeOnDrop`, so
-/// the constraint fires on exactly the bug shape this guard protects against.
-/// The function is never called — only type-checked.
-const _: fn(&RuntimeCredentialMaterialCacheEntry) = |entry| {
-    fn require_zeroize_on_drop<T: ?Sized + secrecy::zeroize::ZeroizeOnDrop>(_: &T) {}
-    if let Some(value) = entry.value.as_ref() {
-        require_zeroize_on_drop(value);
-    }
-};
